move menu to proper folder
